requirement.txt
absl-py==0.9.0
appdirs==1.4.3
appnope==0.1.0
astroid==2.4.0
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.4
backcall==0.1.0
black==20.8b1
bleach==3.1.1
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
chunker==1.1
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.17
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
dnspython==1.16.0
entrypoints==0.3
filelock==3.0.12
fire==0.3.1
future==0.18.2
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gym==0.17.3
h5py==2.10.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
imutils==0.5.3
ipydatawidgets==4.0.1
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipyvolume==0.5.2
ipywebrtc==0.5.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonconfig-tool==1.5
jsonlib-python3==1.6.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras-Preprocessing==1.1.2
keyring==21.2.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
llvmlite==0.32.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
messages==0.5.0
mistune==0.8.4
msgpack==1.0.0
msgpack-numpy==0.4.5
mypy-extensions==0.4.3
mysql-connector-python==8.0.19
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.49.1
numpy==1.19.4
oauthlib==3.1.0
observable==1.0.3
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.6.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycodestyle==2.6.0
pyglet==1.5.0
Pygments==2.5.2
pylint==2.5.0
pyparsing==2.4.7
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyrsistent==0.15.7
pyserial==3.4
pystan==2.19.1.1
python-box==4.2.3
python-constraint==1.4.0
python-dateutil==2.8.1
pythreejs==2.1.1
pytz==2019.3
pyudev==0.22.0
PyWavelets==1.1.1
pyzmq==19.0.0
qtconsole==4.6.0
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
slidingwindow==0.0.13
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorpack @ git+https://github.com/ppwwyyxx/tensorpack.git@ea525564a7edaba6a6c9077b1836988db1860abd
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.5.11
toml==0.10.1
tornado==6.0.3
tqdm==4.46.0
traitlets==4.3.3
traittypes==0.2.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
validus==0.3.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yellowbrick==1.1
